+++ packages/plugin-rsc/examples/starter-cf-single/src/root.tsx
@@ -52,8 +52,8 @@ function App() {
         </li>
         <li>
           Visit{' '}
-          <a href="/?__rsc" target="_blank">
-            <code>/?__rsc</code>
+          <a href="./_.rsc" target="_blank">
+            <code>/_.rsc</code>
           </a>{' '}
           to view RSC stream payload.
         </li>