+++ packages/plugin-rsc/src/react/browser.ts
@@ -54,8 +54,11 @@ export function findSourceMapURL(
   filename: string,
   environmentName: string,
 ): string | null {
-  // TODO: respect config.server.origin and config.base?
-  const url = new URL('/__vite_rsc_findSourceMapURL', window.location.origin)
+  // TODO: respect config.base?
+  const url = new URL(
+    /* @vite-ignore */ '/__vite_rsc_findSourceMapURL',
+    import.meta.url,
+  )
   url.searchParams.set('filename', filename)
   url.searchParams.set('environmentName', environmentName)
   return url.toString()